CL-643: Move the missing impls into the general API pt.2
