+++ core/src/main/java/com/gentics/mesh/core/data/job/impl/JobImpl.java
@@ -58,63 +58,6 @@ public abstract class JobImpl extends AbstractMeshCoreVertex<JobResponse> implem
 	}
 
 	@Override
-	public JobResponse transformToRestSync(InternalActionContext ac, int level, String... languageTags) {
-		JobResponse response = new JobResponse();
-		response.setUuid(getUuid());
-
-		HibUser creator = getCreator();
-		if (creator != null) {
-			response.setCreator(creator.transformToReference());
-		} else {
-			//log.error("The object {" + getClass().getSimpleName() + "} with uuid {" + getUuid() + "} has no creator. Omitting creator field");
-		}
-
-		String date = getCreationDate();
-		response.setCreated(date);
-		response.setErrorMessage(getErrorMessage());
-		response.setErrorDetail(getErrorDetail());
-		response.setType(getType());
-		response.setStatus(getStatus());
-		response.setStopDate(getStopDate());
-		response.setStartDate(getStartDate());
-		response.setCompletionCount(getCompletionCount());
-		response.setNodeName(getNodeName());
-
-		JobWarningList warnings = getWarnings();
-		if (warnings != null) {
-			response.setWarnings(warnings.getData());
-		}
-
-		Map<String, String> props = response.getProperties();
-		HibBranch branch = getBranch();
-		if (branch != null) {
-			props.put("branchName", branch.getName());
-			props.put("branchUuid", branch.getUuid());
-		} else {
-			log.debug("No referenced branch found.");
-		}
-
-		HibSchemaVersion toSchema = getToSchemaVersion();
-		if (toSchema != null) {
-			HibSchema container = toSchema.getSchemaContainer();
-			props.put("schemaName", container.getName());
-			props.put("schemaUuid", container.getUuid());
-			props.put("fromVersion", getFromSchemaVersion().getVersion());
-			props.put("toVersion", toSchema.getVersion());
-		}
-
-		HibMicroschemaVersion toMicroschema = getToMicroschemaVersion();
-		if (toMicroschema != null) {
-			HibMicroschema container = toMicroschema.getSchemaContainer();
-			props.put("microschemaName", container.getName());
-			props.put("microschemaUuid", container.getUuid());
-			props.put("fromVersion", getFromMicroschemaVersion().getVersion());
-			props.put("toVersion", toMicroschema.getVersion());
-		}
-		return response;
-	}
-
-	@Override
 	public JobWarningList getWarnings() {
 		String json = property(WARNING_PROPERTY_KEY);
 		if (json == null) {
