+++ mdm/api/src/main/java/com/gentics/mesh/core/data/group/HibGroup.java
@@ -9,7 +9,9 @@ import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.TypeInfo;
 import com.gentics.mesh.core.data.HibBucketableElement;
 import com.gentics.mesh.core.data.HibNamedElement;
+import com.gentics.mesh.core.data.dao.GroupDao;
 import com.gentics.mesh.core.data.user.HibUserTracking;
+import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.rest.group.GroupReference;
 import com.gentics.mesh.core.rest.group.GroupResponse;
 import com.gentics.mesh.handler.VersionUtils;
@@ -56,5 +58,11 @@ public interface HibGroup extends HibNamedElement<GroupResponse>, HibUserTrackin
 	@Override
 	default String getAPIPath(InternalActionContext ac) {
 		return VersionUtils.baseRoute(ac) + "/groups/" + getUuid();
+	}
+
+	@Override
+	default GroupResponse transformToRestSync(InternalActionContext ac, int level, String... languageTags) {
+		GroupDao groupDao = Tx.get().groupDao();
+		return groupDao.transformToRestSync(this, ac, level, languageTags);
 	}
 }
