+++ mdm/api/src/main/java/com/gentics/mesh/core/data/schema/HibSchema.java
@@ -1,7 +1,5 @@
 package com.gentics.mesh.core.data.schema;
 
-import static com.gentics.mesh.MeshVersion.CURRENT_API_BASE_PATH;
-
 import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibBucketableElement;
 import com.gentics.mesh.core.rest.schema.SchemaReference;
