+++ mdm/api/src/main/java/com/gentics/mesh/core/data/tagfamily/HibTagFamily.java
@@ -5,8 +5,10 @@ import static com.gentics.mesh.util.URIUtils.encodeSegment;
 import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibBucketableElement;
 import com.gentics.mesh.core.data.HibCoreElement;
+import com.gentics.mesh.core.data.dao.TagFamilyDao;
 import com.gentics.mesh.core.data.project.HibProject;
 import com.gentics.mesh.core.data.user.HibUserTracking;
+import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.rest.tag.TagFamilyReference;
 import com.gentics.mesh.core.rest.tag.TagFamilyResponse;
 import com.gentics.mesh.handler.VersionUtils;
@@ -75,5 +77,12 @@ public interface HibTagFamily extends HibCoreElement<TagFamilyResponse>, HibUser
 	@Override
 	default String getAPIPath(InternalActionContext ac) {
 		return VersionUtils.baseRoute(ac) + "/" + encodeSegment(getProject().getName()) + "/tagFamilies/" + getUuid();
+	}
+
+	@Override
+	@Deprecated
+	default TagFamilyResponse transformToRestSync(InternalActionContext ac, int level, String... languageTags) {
+		TagFamilyDao tagFamilyDao = Tx.get().tagFamilyDao();
+		return tagFamilyDao.transformToRestSync(this, ac, level, languageTags);
 	}
 }
